refactor(setup): extract shared tool-probing helpers; drop dead adapter
The `CORE_TOOLS` list and the known-config-plus-PATH-fallback probe were
duplicated across ProbeEnvironmentTool and VerifySetupTool, forcing
every review pass to remember to update both sides. Hoist them into a
single `toolProbing.ts` module and make both tools import from it, so
the LaTeX toolchain list and the probe semantics have one source of
truth.

Drive-by cleanups for the two unused exports Cursor Bugbot flagged
(Low severity):
- `SetupCommandAdapter.hasCommand` was declared and wired in
  extension.ts but never called by any setup tool — drop it.
- `setupAssistantCommands` was re-exported from the `commands/setup`
  barrel but only consumed internally; drop the barrel entry.

No behavior change.
